# contracts/openapi/bff-v1.yaml
openapi: 3.0.3
info:
  title: UnoArena Realtime Gateway / BFF
  version: 1.0.0
  description: |
    Sole external HTTP/SSE surface for UnoArena. Clients submit versioned command
    envelopes, fetch reconnect/projection snapshots, and subscribe to SSE streams.
    Microservices are not publicly reachable. For this implementation the repo-owned
    simple CLI is the sole client/test interface; a graphical UI is deferred and must
    continue to use only this BFF boundary. Rejected commands return `status=rejected`
    with a structured operational/security audit record only (never domain events or
    Game Integrity entries). Unknown or evicted `Last-Event-ID` values return HTTP 409
    `snapshot_required` so clients resync via the snapshot routes before resuming SSE.
servers:
  - url: http://localhost:8080
    description: Local compose BFF (gateway)
...
paths:
...
  /v1/analytics/public:
    get:
      tags: [reads]
      summary: Public analytics read model (proxied)
      operationId: getPublicAnalytics
      parameters:
        - $ref: "#/components/parameters/CorrelationIdHeader"
      responses:
        "200":
          description: Derived public analytics
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "429":
          $ref: "#/components/responses/Error"
        "502":
          description: Analytics upstream failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
components:
  securitySchemes:
    SessionBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
      description: |
        Identity-issued session token via BFF login. Required for player/command/control
        routes. Optional on spectator routes; when present, invalid tokens yield 401.
  parameters:
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
...
  responses:
    Error:
      description: Structured error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorBody"
...
  schemas:
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string
        commandId:
          type: string
...